refactor(blorktools): variable cleanup
